infrastructure/requirements.txt
aws_cdk.core>=1.85.0,<2.0
aws_cdk.aws_dynamodb>=1.85.0,<2.0
aws_cdk.aws_sqs>=1.85.0,<2.0
chalice[cdk] ~= 1.22.3
